Organize the tests into detailed categorized files, and removed duplicated test.
